+++ Realm/RLMRealm.mm
@@ -447,6 +447,12 @@ REALM_NOINLINE void RLMRealmTranslateException(NSError **error) {
         }
     }
 
+    if (configuration.seedFilePath && ![RLMRealm fileExistsForConfiguration:configuration]) {
+        [[NSFileManager defaultManager] copyItemAtURL:configuration.seedFilePath
+                                                toURL:[configuration fileURL]
+                                                error:error];
+    }
+
     configuration = [configuration copy];
     Realm::Config& config = configuration.config;