+++ Realm/Tests/RealmTests.mm
@@ -1839,7 +1839,7 @@
     XCTAssertFalse([realm writeCopyToURL:RLMTestRealmURL() encryptionKey:nil error:&writeError]);
     RLMValidateRealmError(writeError, RLMErrorFilePermissionDenied, expectedError, expectedUnderlying);
 
-    // Test writeCopyWithConfiguration
+    // Test writeCopyForConfiguration
     RLMRealmConfiguration *configuration = [RLMRealmConfiguration defaultConfiguration];
     configuration.fileURL = RLMTestRealmURL();
     writeError = nil;
@@ -1896,9 +1896,9 @@
     }];
 }
 
-#pragma mark - Write Copy with Configuration
+#pragma mark - Write Copy For Configuration
 
-- (void)testWriteCopyWithConfiguration
+- (void)testWriteCopyForConfiguration
 {
     RLMRealmConfiguration *configuration = [RLMRealmConfiguration defaultConfiguration];
     configuration.fileURL = RLMTestRealmURL();